refactor: Migrate, cleanup, and document `helper.js`, `logging.js`, and `title-rename.js` to typescript
