+++ package.json
@@ -36,6 +36,7 @@
     "@nextcloud/moment": "^1.2.2",
     "@nextcloud/router": "^2.2.0",
     "@nextcloud/vue": "^8.3.0",
+    "axios": "^1.6.5",
     "caret-pos": "^2.0.0",
     "fuse.js": "^7.0.0",
     "linkifyjs": "^4.1.1",
@@ -57,6 +58,7 @@
     "@nextcloud/browserslist-config": "^3.0.0",
     "@nextcloud/stylelint-config": "^2.3.1",
     "@nextcloud/webpack-vue-config": "^6.0.0",
+    "@types/jest": "^29.5.11",
     "@typescript-eslint/eslint-plugin": "^6.18.0",
     "@typescript-eslint/parser": "^6.18.0",
     "@vue/cli-plugin-typescript": "~5.0.8",
