--- src/tests/.eslintrc.yml
@@ -1,2 +1,0 @@
-env:
-    jest: true
