+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
 	"include": ["src/**/*.ts", "src/**/*.js", "src/**/*.vue"],
-	"exclude": ["node_modules"],
+	"exclude": ["node_modules", "js"],
 	"compilerOptions": {
 		"lib": [
 			"dom",
